alt/ericvw: Change startup lighting to solid blueish that is reactive
This is the lighting effect I keep gravitating to each day.  For now,
encode more-or-less what I would do if I was setting this up through the
function layer.  Later on, I should consider hard-coding the HSV values
so that it remains stable in case the defaults were to ever change.