alt/ericvw: Add personal daily driving layer
Having the Caps Lock key producing Esc is easier for me to reach when
using `vim`, after having used it with my iKBC Poker II programming and
software mapping on macOS.

Additionally, Caps Lock on macOS requires a delay for activating, which
makes it unusable for gaming.  This opens of the physical Esc key for
emitting Grave, now that emitting Esc is still possible.

Muscle memory prevails for wanting the Grave key above the Tab key and
left of the 1 key.